Stammdaten: Album, Kategorie, Kategoriewert fertig

+++ FamilieLaissSQLServer/dbo/Tables/Facet_Group.sql
@@ -4,10 +4,12 @@
     [Name_German]      NVARCHAR (70)      NOT NULL,
     [Name_English]     NVARCHAR (70)      NOT NULL,
     [Can_Delete]       BIT                NOT NULL,
-    [Facet_Value_Type] INT                NOT NULL,
+    [Facet_Value_Type] TINYINT            NOT NULL,
     [DDL_Create]       DATETIMEOFFSET (7) NOT NULL,
     PRIMARY KEY CLUSTERED ([ID] ASC)
 );
+
+
 
 
 GO
